+++ CHANGELOG.md
@@ -9,10 +9,12 @@ All notable changes to this project will be documented in this file.
 - Include chart name when installing with a custom release name ([#97])
 - Pinned MinIO version for tests ([#100])
 - `operator-rs` `0.21.0` → `0.22.0` ([#102]).
+- Added owner-reference to pod templates ([#104])
 
 [#97]: https://github.com/stackabletech/spark-k8s-operator/pull/92
 [#100]: https://github.com/stackabletech/spark-k8s-operator/pull/100
 [#102]: https://github.com/stackabletech/spark-k8s-operator/pull/102
+[#104]: https://github.com/stackabletech/spark-k8s-operator/pull/104
 
 ## [0.3.0] - 2022-06-30